test(infra): extract Unit-traited tests from Integration collections
`Backend Fast (build + unit)` runs with `--filter "Category=Unit"` and NO
Docker setup. 5 tests with `[Trait("Category", TestCategories.Unit)]` lived
inside classes bound to a Testcontainers collection (`Integration-GroupC`,
`Integration-GroupD`). xUnit instantiates the collection fixture for ALL
tests in the collection — including the Unit-traited ones — so
`SharedTestcontainersFixture.InitializeAsync()` ran and failed with
`DockerUnavailableException` on the Unit job.

Result: those 5 tests have been broken on `main-dev` since at least PR #1863
(merged 2026-06-03 14:54), but the failures were not attributed to the
introducing PR because subsequent backend PRs skipped Backend Fast via the
Detect Changes filter.

This PR extracts the 5 affected tests into standalone Unit classes (no
Collection attribute, no fixture):

- `PuertoRicoGoldenSeederUnitTests` (new) — 4 tests:
  - LoadEmbeddedFixture_ParsesAtLeastFiftyClaims
  - LoadEmbeddedFixture_AllSixSectionsRepresented
  - LoadEmbeddedBggTagsForTest_ReturnsParsedFixture
  - SeedAsync_WhenSkipFlagTrue_ReturnsImmediatelyAndSendsNothing
  Each is a static / mock-only invocation — no DbContext needed.

- `DomainEventBroadcastInterceptorOrderingTests` (new) — 1 test:
  - Publishes_AfterCommit_NotBefore
  Uses EF Core InMemory only; the post-commit ordering property is an EF
  Core contract independent of the database provider.

The original `PuertoRicoGoldenSeederTests` and
`DomainEventBroadcastInterceptorTests` retain their Integration tests with
the Testcontainers collection — that wiring is correct for tests that
actually touch Postgres.

5/5 extracted Unit tests passing locally (Release, no Docker, ~2s).

Issue #1873 review follow-up.
