+++ apps/api/tests/Api.Tests/Infrastructure/EventBroadcasting/DomainEventBroadcastInterceptorTests.cs
@@ -184,98 +184,9 @@ public sealed class DomainEventBroadcastInterceptorTests : IAsyncLifetime
     }
 
     // -------------------------------------------------------------------------
-    // Test 3: Publishes_AfterCommit_NotBefore
+    // Test 3 (Publishes_AfterCommit_NotBefore) has been moved to
+    // DomainEventBroadcastInterceptorOrderingTests so it does not trigger the
+    // Integration-GroupD Testcontainers fixture on the Backend Fast CI job
+    // (no Docker). Issue #1873 review follow-up.
     // -------------------------------------------------------------------------
-
-    /// <summary>
-    /// Verifies ordering: Publish is called AFTER <c>SaveChangesAsync</c> completes
-    /// (i.e., from <c>SavedChangesAsync</c>, not from <c>SavingChangesAsync</c>).
-    ///
-    /// Uses EF InMemory so the test does not require Testcontainers infra — the
-    /// ordering property is an EF Core contract (SavedChangesAsync fires post-SQL),
-    /// independent of the database provider.
-    ///
-    /// Timing instrumentation: <c>saveChangesStarted</c> is flipped to <c>true</c>
-    /// just before <c>SaveChangesAsync</c> is called. The Publish callback checks
-    /// this flag and records whether Publish fired while <c>SaveChangesAsync</c>
-    /// was still executing (i.e., before it returned control to the caller). If the
-    /// interceptor were to call Publish from <c>SavingChangesAsync</c> (pre-commit),
-    /// the <c>publishedBeforeSaveCompleted</c> flag would be set to <c>true</c>,
-    /// causing the assertion to fail.
-    /// </summary>
-    [Fact]
-    [Trait("Category", TestCategories.Unit)]
-    public async Task Publishes_AfterCommit_NotBefore()
-    {
-        // Arrange
-        var saveChangesStarted = false;
-        var publishedBeforeSaveCompleted = false;
-        var publishWasCalled = false;
-
-        var broadcasterMock = new Mock<IEventBroadcaster>();
-        broadcasterMock
-            .Setup(b => b.Publish(It.IsAny<DomainEventDto>()))
-            .Callback<DomainEventDto>(_ =>
-            {
-                // If saveChangesStarted is true but SaveChangesAsync has not yet returned,
-                // we are inside the interceptor call chain. A correct implementation calls
-                // Publish from SavedChangesAsync (which runs before SaveChangesAsync returns),
-                // so this flag will be true here — that is expected and correct.
-                // What we must PREVENT is Publish being called before saveChangesStarted,
-                // which would mean it fired outside any SaveChanges call entirely.
-                if (!saveChangesStarted)
-                {
-                    // Publish fired before SaveChangesAsync was even entered — regression.
-                    publishedBeforeSaveCompleted = true;
-                }
-                publishWasCalled = true;
-            });
-
-        var interceptor = new DomainEventBroadcastInterceptor(broadcasterMock.Object);
-
-        var options = new DbContextOptionsBuilder<MeepleAiDbContext>()
-            .UseInMemoryDatabase($"timing-test-{Guid.NewGuid():N}")
-            .ConfigureWarnings(w =>
-            {
-                w.Ignore(Microsoft.EntityFrameworkCore.Diagnostics.RelationalEventId.PendingModelChangesWarning);
-                w.Ignore(Microsoft.EntityFrameworkCore.Diagnostics.InMemoryEventId.TransactionIgnoredWarning);
-            })
-            .AddInterceptors(interceptor)
-            .Options;
-
-        var mockMediator = new Mock<IMediator>();
-        var mockEventCollector = new Mock<IDomainEventCollector>();
-        mockEventCollector.Setup(e => e.PeekEvents()).Returns(Array.Empty<IDomainEvent>());
-        mockEventCollector.Setup(e => e.Clear());
-
-        await using var ctx = new MeepleAiDbContext(options, mockMediator.Object, mockEventCollector.Object);
-
-        var entity = MakeEventLog();
-        ctx.DomainEventLogs.Add(entity);
-
-        // Verify Publish has NOT been called before SaveChanges starts
-        publishWasCalled.Should().BeFalse("Publish must not fire before SaveChangesAsync is called");
-
-        // Act — EF Core fires SavingChangesAsync (before SQL) then SavedChangesAsync (after SQL).
-        // Our interceptor publishes in SavedChangesAsync, so publishWasCalled becomes true
-        // during the SaveChangesAsync call but AFTER the underlying SQL/InMemory write.
-        saveChangesStarted = true;
-        await ctx.SaveChangesAsync();
-
-        // After SaveChangesAsync returns, SavedChangesAsync has already completed → Publish was invoked.
-        publishWasCalled.Should().BeTrue(
-            "Publish must have been called (from SavedChangesAsync, post-commit) by the time SaveChangesAsync returns");
-
-        // Real regression guard: if Publish fired before saveChangesStarted was set,
-        // it means the interceptor called Publish outside the SaveChanges pipeline entirely.
-        publishedBeforeSaveCompleted.Should().BeFalse(
-            "Publish must not fire before SaveChangesAsync is entered — it must be called from SavedChangesAsync (post-commit), not SavingChangesAsync (pre-commit)");
-
-        // Validate the published DTO maps correctly from the entity
-        broadcasterMock.Verify(b => b.Publish(It.Is<DomainEventDto>(d =>
-            d.Id == entity.Id &&
-            d.EventId == entity.EventId &&
-            d.EventType == entity.EventType &&
-            d.PayloadJson == entity.PayloadJson)), Times.Once);
-    }
 }
