+++ apps/api/tests/Api.Tests/SharedGameCatalog/Seeding/PuertoRicoGoldenSeederTests.cs
@@ -61,76 +61,11 @@ public sealed class PuertoRicoGoldenSeederTests : IAsyncLifetime
     }
 
     // ============================================================
-    // Embedded resource sanity (unit-style, no DB calls reach Postgres)
+    // Unit-only tests (LoadEmbedded* + SeedAsync_WhenSkipFlagTrue_*)
+    // have been moved to PuertoRicoGoldenSeederUnitTests so they no
+    // longer trigger the Integration-GroupC Testcontainers fixture on
+    // the Backend Fast CI job. Issue #1873 review follow-up.
     // ============================================================
-
-    [Fact]
-    [Trait("Category", TestCategories.Unit)]
-    public void LoadEmbeddedFixture_ParsesAtLeastFiftyClaims()
-    {
-        var claims = PuertoRicoGoldenSeeder.LoadEmbeddedClaimsForTest();
-
-        claims.Should().NotBeNull();
-        claims.Count.Should().BeGreaterThanOrEqualTo(
-            50,
-            "Sprint 2 spec requires at least 50 curated Puerto Rico claims (Task 1)");
-
-        claims.Should().AllSatisfy(c =>
-        {
-            c.Statement.Should().NotBeNullOrWhiteSpace();
-            c.SourceQuote.Should().NotBeNullOrWhiteSpace();
-            c.ExpectedPage.Should().BeGreaterThanOrEqualTo(1);
-            // Section is an int 0..5; the enum has 6 members.
-            c.Section.Should().BeInRange(0, 5);
-        });
-    }
-
-    [Fact]
-    [Trait("Category", TestCategories.Unit)]
-    public void LoadEmbeddedFixture_AllSixSectionsRepresented()
-    {
-        var claims = PuertoRicoGoldenSeeder.LoadEmbeddedClaimsForTest();
-
-        var sections = claims.Select(c => c.Section).Distinct().OrderBy(s => s).ToArray();
-        sections.Should().BeEquivalentTo(
-            new[] { 0, 1, 2, 3, 4, 5 },
-            "Task 1 fixture audit guarantees ≥1 claim per MechanicSection");
-    }
-
-    // ============================================================
-    // Skip flag (unit-style — never touches DbContext)
-    // ============================================================
-
-    [Fact]
-    [Trait("Category", TestCategories.Unit)]
-    public async Task SeedAsync_WhenSkipFlagTrue_ReturnsImmediatelyAndSendsNothing()
-    {
-        var config = new ConfigurationBuilder()
-            .AddInMemoryCollection(new Dictionary<string, string?>
-            {
-                ["SKIP_PUERTO_RICO_GOLDEN_SEED"] = "true"
-            })
-            .Build();
-
-        var mediator = new Mock<IMediator>(MockBehavior.Strict);
-        var services = new ServiceCollection()
-            .AddSingleton<IConfiguration>(config)
-            .AddSingleton<IMediator>(mediator.Object)
-            .BuildServiceProvider();
-
-        var context = new SeedContext(
-            Profile: SeedProfile.Dev,
-            DbContext: null!, // Skip path must NOT dereference the DbContext.
-            Services: services,
-            Logger: NullLogger.Instance,
-            SystemUserId: Guid.NewGuid());
-
-        var seeder = new PuertoRicoGoldenSeeder();
-
-        await seeder.SeedAsync(context, default);
-
-        mediator.VerifyNoOtherCalls();
-    }
 
     // ============================================================
     // Integration: real Postgres + EF + MediatR mock
@@ -373,26 +308,10 @@ public sealed class PuertoRicoGoldenSeederTests : IAsyncLifetime
     }
 
     // ============================================================
-    // Task 4: BGG tags seeder
+    // Task 4: BGG tags seeder (Integration tests only;
+    // LoadEmbeddedBggTagsForTest_ReturnsParsedFixture lives in
+    // PuertoRicoGoldenSeederUnitTests — see header comment above.)
     // ============================================================
-
-    [Fact]
-    [Trait("Category", TestCategories.Unit)]
-    public void LoadEmbeddedBggTagsForTest_ReturnsParsedFixture()
-    {
-        var tags = PuertoRicoGoldenSeeder.LoadEmbeddedBggTagsForTest();
-
-        tags.Should().NotBeNull();
-        tags.Count.Should().BeGreaterThanOrEqualTo(
-            10,
-            "Sprint 2 Task 4 fixture must contain at least 10 BGG tags for Puerto Rico");
-
-        tags.Should().AllSatisfy(t =>
-        {
-            t.Name.Should().NotBeNullOrWhiteSpace();
-            t.Category.Should().NotBeNullOrWhiteSpace();
-        });
-    }
 
     [Fact]
     [Trait("Category", TestCategories.Integration)]
